+++ setup.cfg
@@ -5,6 +5,9 @@ filterwarnings =
     error
     ignore::DeprecationWarning:distutils
     ignore::DeprecationWarning:Cython
+    # for python 3.10+: mypy still relies on distutils on windows. We hence ignore those warnings
+    ignore:The distutils package is deprecated and slated for removal in Python 3.12:DeprecationWarning
+    ignore:The distutils.sysconfig module is deprecated, use sysconfig instead:DeprecationWarning
 
 [flake8]
 max-line-length = 120
